requirements.txt
Flask==2.1.2
pendulum==2.1.2
python-dotenv==0.20.0
gunicorn==20.1.0